fi_read_file function moved to OSD
due to POSIX API used in fi_read_file function declared
as deprecated in Windows created more 'native' implementation
of fi_read_file function for windows platform
